add image loading to log file
